+++ src/Plutarch/Extra/Map.hs
@@ -14,6 +14,7 @@
 module Plutarch.Extra.Map (
     -- * Lookup
     ptryLookup,
+    plookupGe,
 
     -- * Comparisons
     pkeysEqual,
@@ -50,6 +51,38 @@ import Plutarch.Api.V1.AssocMap (
 import Plutarch.Builtin (ppairDataBuiltin)
 import Plutarch.Extra.Maybe (passertPJust)
 import qualified Plutarch.List as PList
+
+{- | As 'plookup', but also yields the portion of the 'PMap' whose keys are
+ greater than the target if the search is successful.
+
+ @since 3.8.0
+-}
+plookupGe ::
+    forall (k :: S -> Type) (v :: S -> Type) (s :: S).
+    (PIsData k, POrd k) =>
+    Term s (k :--> PMap 'Sorted k v :--> PMaybe (PPair (PAsData v) (PMap 'Sorted k v)))
+plookupGe = phoistAcyclic $
+    plam $ \needle ->
+        pfix #$ plam $ \self xs ->
+            pelimList (go needle self) (pcon PNothing) . pto $ xs
+  where
+    go ::
+        forall (s' :: S).
+        Term s' k ->
+        Term s' (PMap 'Sorted k v :--> PMaybe (PPair (PAsData v) (PMap 'Sorted k v))) ->
+        Term s' (PBuiltinPair (PAsData k) (PAsData v)) ->
+        Term s' (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))) ->
+        Term s' (PMaybe (PPair (PAsData v) (PMap 'Sorted k v)))
+    go needle self h t =
+        let current = pfromData $ pfstBuiltin # h
+         in pif
+                (needle #== current)
+                (pcon . PJust . pcon . PPair (psndBuiltin # h) . pcon . PMap $ t)
+                ( pif
+                    (needle #< current)
+                    (pcon PNothing)
+                    (self # (pcon . PMap $ t))
+                )
 
 {- | If a value exists at the specified key, apply the function argument to it;
  otherwise, do nothing.